Move Django formatter patches to their own file
